create Interfaces.ts to store all interfaces
- Create Utils folder to hold helper files
- Move interfaces from CustomerStore.tsx to Interfaces.ts
